feat: Cross iframe is enabled by default
