+++ .changeset/sour-coats-float.md
@@ -1,0 +1,8 @@
+---
+'@open-editor/webpack': patch
+'@open-editor/client': patch
+'@open-editor/rollup': patch
+'@open-editor/vite': patch
+---
+
+Cross iframe is enabled by default
