+++ packages/client/src/ui/TreeUI.tsx
@@ -1,4 +1,5 @@
 import { addClass, removeClass, getHtml, replaceChildren, applyStyle } from '../utils/dom';
+import { inspectorState } from '../inspector/inspectorState';
 import { treeCloseBridge, openEditorBridge, treeOpenBridge } from '../bridge';
 import { getOptions } from '../options';
 import { type CodeSource, type CodeSourceMeta } from '../resolve';
@@ -15,6 +16,8 @@ export function TreeUI() {
   };
 
   treeOpenBridge.on((source) => {
+    inspectorState.isTreeOpen = true;
+
     applyStyle(state.root, {
       display: 'block',
     });
@@ -25,6 +28,8 @@ export function TreeUI() {
   });
 
   treeCloseBridge.on(() => {
+    inspectorState.isTreeOpen = false;
+
     applyStyle(state.root, {
       display: 'none',
     });
