+++ packages/webpack/README.md
@@ -70,7 +70,7 @@ interface Options {
    *
    * It only takes effect when the top window and iframe window have the same origin.
    *
-   * @default false
+   * @default true
    */
   crossIframe?: boolean;
   /**
